requirements.txt
certifi==2023.7.22
pendulum==2.1.2
pymongo==4.4.1
python-dotenv==1.0.0
python-telegram-bot==20.4
selenium==4.10.0
webdriver-manager==4.0.1
